requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
cloudinary==1.28.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.0.1
django-ckeditor==6.2.0
django-heroku==0.0.0
django-js-asset==1.2.2
gunicorn==20.1.0
idna==3.3
Pillow==9.0.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
whitenoise==5.3.0
